pyproject.toml
[project]
name = "financeapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "jpype1>=1.5.2",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "streamlit>=1.45.0",
    "tabula-py>=2.10.0",
]
